+++ src/vue.ts
@@ -2,7 +2,7 @@ import {vue} from '@codemirror/lang-vue';
 import {htmlLanguage, htmlCompletionSource} from '@codemirror/lang-html';
 import {javascript} from '@codemirror/lang-javascript';
 import {LanguageSupport} from '@codemirror/language';
-import {cssCompletion} from './css.js';
+import {cssCompletion, markLinkPlugin} from './css.js';
 import {jsCompletion, markGlobalsAndDocTagPlugin} from './javascript.js';
 import type {CodeMirror6} from './codemirror';
 
@@ -13,5 +13,6 @@ export default (_?: unknown, cm?: CodeMirror6): LanguageSupport => vue({
 		jsCompletion,
 		cssCompletion(),
 		markGlobalsAndDocTagPlugin(cm),
+		markLinkPlugin(cm),
 	]),
 });