+++ test/src/css.test.ts
@@ -1,7 +1,12 @@
 import {pathToFileURL} from 'url';
+import * as assert from 'assert';
+import {syntaxTree} from '@codemirror/language';
+import {html} from '@codemirror/lang-html';
+import {vue} from '@codemirror/lang-vue';
 import {describe, it} from '@bhsd/test-util/mocha';
-import css from '../../dist/css.js';
-import {autocompletionTest, createState} from './util.js';
+import css, {markLink} from '../../dist/css.js';
+import {autocompletionTest, createState, convertFullRangeSet, filterFromRangeSet} from './util.js';
+import type {LanguageSupport} from '@codemirror/language';
 import type {CompletionSource, CompletionResult} from '@codemirror/autocomplete';
 import type {Dialect} from '../../dist/codemirror';
 
@@ -32,10 +37,26 @@ Object.assign(globalThis, {
 
 const mockTest = (dialect?: Dialect): (doc: string, result: CompletionResult | null) => Promise<void> => {
 	const lang = css(dialect),
-		state = createState('*', lang),
-		[source] = state.languageDataAt<CompletionSource>('autocomplete', 0);
+		[source] = createState('*', lang).languageDataAt<CompletionSource>('autocomplete', 0);
 	// eslint-disable-next-line require-unicode-regexp, regexp/no-empty-alternative
 	return autocompletionTest(source!, lang, /^(\w[\w-]*|-\w[\w-]*|)$/);
+};
+
+const linkTest = (doc: string, links: [number, number][], lang = css(undefined)): void => {
+	const state = createState(doc, lang),
+		{length} = doc,
+		set = markLink(syntaxTree(state), [{from: 0, to: length}], state);
+	assert.deepStrictEqual(filterFromRangeSet(convertFullRangeSet(set, length), 'cm-link'), links);
+};
+
+const sublangTest = (name: string, lang: LanguageSupport): void => {
+	it(`CSS nested in ${name}`, () => {
+		linkTest(
+			'<style>/* See https://www.w3.org/TR/css-flexbox-1/#min-size-auto */</style>',
+			[[14, 64]],
+			lang,
+		);
+	});
 };
 
 describe('CSS autocompletion', () => {
@@ -57,4 +78,12 @@ describe('CSS autocompletion', () => {
 			options: [{label: '-webkit-user-select', type: 'property', apply: '-webkit-user-select: '}],
 		});
 	});
+});
+
+describe('links in CSS comments', () => {
+	it('CSS', () => {
+		linkTest('/* See https://www.w3.org/TR/css-flexbox-1/#min-size-auto */', [[7, 57]]);
+	});
+	sublangTest('HTML', html());
+	sublangTest('Vue', vue());
 });